requirements.txt
torch
transformers
accelerate
bitsandbytes
llama-cpp-python
streamlit
huggingface-hub

